+++ ArUCO/src/calibration.cpp
@@ -5,7 +5,7 @@
 #include "calibration.h"
 
 
-void boardToPng(const std::string &path, uint8_t markersX, uint8_t markersY, uint8_t markerLen, uint8_t markerSep) {
+void boardToPng(const std::string &path, int markersX, int markersY, int markerLen, int markerSep) {
     int margins = markerSep;
     int borderBits = 1;
 
@@ -27,7 +27,7 @@ void boardToPng(const std::string &path, uint8_t markersX, uint8_t markersY, uin
 }
 
 
-uint8_t calibrate(Camera camera, uint8_t markersX, uint8_t markersY, uint8_t markerLen, uint8_t markerSep) {
+uint8_t calibrate(Camera camera, uint8_t markersX, uint8_t markersY, float markerLen, float markerSep) {
     cv::Size imgSize;
     cv::Mat img, imgCopy;
